Extract EnvVarGuard for tests
