+++ tests/cucumber.rs
@@ -1,4 +1,5 @@
 mod steps;
+mod support;
 use cucumber::World as _;
 use steps::{CliWorld, ClientWorld, CommentWorld, ConfigWorld, ListenerWorld, WorkerWorld};
