+++ tests/support/mod.rs
@@ -1,0 +1,3 @@
+//! Support utilities shared by tests.
+
+pub mod env_guard;
